Refactor to PromptSuite: rename package, update imports and structure
