--- multipromptify_tasks/__init__.py
@@ -1,6 +1,0 @@
-"""
-MultiPromptify Tasks Package
-A collection of tasks for generating prompt variations using the MultiPromptify framework.
-"""
-
-__version__ = "1.0.0"
