+++ promptsuite_tasks/__init__.py
@@ -1,0 +1,6 @@
+"""
+PromptSuite Tasks Package
+A collection of tasks for generating prompt variations using the PromptSuite framework.
+"""
+
+__version__ = "1.0.0"
