+++ promptsuite_tasks/data_generation/math_task.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """
-Sentiment Analysis Task: Stanford Sentiment Treebank (SST)
-This module provides a class for generating prompt variations for sentiment analysis tasks.
+Math Problems Task: GSM8K
+This module provides a class for generating prompt variations for math problem solving tasks.
 """
 
 from typing import Dict, Any, List
@@ -9,23 +9,23 @@ import argparse
 import pandas as pd
 from pathlib import Path
 import sys
-
+import re
 
 from datasets import load_dataset
-from multipromptify.core.template_keys import (
+from promptsuite.core.template_keys import (
     INSTRUCTION, PROMPT_FORMAT, GOLD_KEY,
     PARAPHRASE_WITH_LLM, FORMAT_STRUCTURE_VARIATION, TYPOS_AND_NOISE_VARIATION,
     INSTRUCTION_VARIATIONS, PROMPT_FORMAT_VARIATIONS, FEW_SHOT_KEY
 )
 from base_task import BaseTask
-from multipromptify_tasks.constants import (
+from promptsuite_tasks.constants import (
     DEFAULT_VARIATIONS_PER_FIELD, DEFAULT_PLATFORM, DEFAULT_MODEL_NAME,
     DEFAULT_MAX_VARIATIONS_PER_ROW, DEFAULT_MAX_ROWS, DEFAULT_RANDOM_SEED
 )
 
 
-class SentimentTask(BaseTask):
-    """Task for generating sentiment analysis prompt variations."""
+class MathTask(BaseTask):
+    """Task for generating math problem solving prompt variations."""
 
     def __init__(self,
                  variations_per_field: int = DEFAULT_VARIATIONS_PER_FIELD,
@@ -35,8 +35,8 @@ class SentimentTask(BaseTask):
                  max_variations_per_row: int = DEFAULT_MAX_VARIATIONS_PER_ROW,
                  random_seed: int = DEFAULT_RANDOM_SEED):
 
-        task_name = "Sentiment Analysis Task: SST"
-        output_filename = "sentiment_sst_variations.json"
+        task_name = "Math Problems Task: GSM8K"
+        output_filename = "math_gsm8k_variations.json"
 
         super().__init__(
             task_name=task_name,
@@ -50,38 +50,35 @@ class SentimentTask(BaseTask):
         )
 
     def load_data(self) -> None:
-        """Load Stanford Sentiment Treebank (SST) dataset from HuggingFace - both train and test splits, combine, and load as one DataFrame."""
-        print("Loading SST train dataset...")
-        train_ds = load_dataset("stanfordnlp/sst", trust_remote_code=True, split="train[:100]")
+        """Load GSM8K dataset from HuggingFace - both train and test splits, combine, and load as one DataFrame."""
+        print("Loading GSM8K train dataset...")
+        train_ds = load_dataset("gsm8k", "main", split="train[:100]")
         train_df = pd.DataFrame(train_ds)
-
-        # Convert labels to continuous sentiment scores (0.0 to 1.0)
-        # SST labels are typically 0-4, we normalize to 0.0-1.0
         train_df['split'] = 'train'
         print(f"✅ Loaded {len(train_df)} train rows")
 
-        print("Loading SST test dataset...")
-        test_ds = load_dataset("stanfordnlp/sst", trust_remote_code=True, split="test[:100]")
+        print("Loading GSM8K test dataset...")
+        test_ds = load_dataset("gsm8k", "main", split="test[:100]")
         test_df = pd.DataFrame(test_ds)
         test_df['split'] = 'test'
         print(f"✅ Loaded {len(test_df)} test rows")
 
-        # Combine
+        # Combine datasets
         df = pd.concat([train_df, test_df], ignore_index=True)
         print(f"✅ Combined total: {len(df)} rows")
 
         self.mp.load_dataframe(df)
-        print("✅ Data post-processed")
+        print("✅ Data loaded")
 
     def get_template(self) -> Dict[str, Any]:
-        """Get template configuration for sentiment analysis task with continuous scoring."""
+        """Get template configuration for math problem solving task."""
         return {
-            INSTRUCTION: "Rate the sentiment of the following text on a scale from 0.0 (most negative) to 1.0 (most positive).",
-            INSTRUCTION_VARIATIONS: [PARAPHRASE_WITH_LLM],
-            PROMPT_FORMAT: "Text: \"{sentence}\"\nSentiment Score: {label}",
+            INSTRUCTION: "Let's think step by step. Solve the following math problem and provide the final numerical answer in the format #### answer.",
+            # INSTRUCTION_VARIATIONS: [PARAPHRASE_WITH_LLM],
+            PROMPT_FORMAT: "Question: {question}\nAnswer: {answer}",
             PROMPT_FORMAT_VARIATIONS: [FORMAT_STRUCTURE_VARIATION],
-            'sentence': [TYPOS_AND_NOISE_VARIATION],
-            GOLD_KEY: 'label',  # The target field for evaluation
+            'question': [TYPOS_AND_NOISE_VARIATION],
+            GOLD_KEY: 'answer',  # The original answer field with full solution and #### format
             FEW_SHOT_KEY: {
                 'count': 3,  # Number of few-shot examples
                 'format': 'random_per_row',  # Random examples per row
@@ -90,17 +87,17 @@ class SentimentTask(BaseTask):
         }
 
 
-def generate_sentiment_variations(variations_per_field, api_platform, model_name, max_rows, max_variations_per_row, random_seed):
-    """Generate variations for sentiment analysis task."""
+def generate_math_variations(variations_per_field, api_platform, model_name, max_rows, max_variations_per_row, random_seed):
+    """Generate variations for math problems task."""
     # Create output directory
-    output_dir = Path(__file__).parent.parent / "generated_data"/ "data" / "sentiment"
+    output_dir = Path(__file__).parent.parent / "tasks_data" / "generated_data" / "math"
     output_dir.mkdir(parents=True, exist_ok=True)
 
-    print("🎯 Processing Sentiment Analysis Task")
+    print("🎯 Processing Math Problems Task")
     print("=" * 50)
 
     try:
-        task = SentimentTask(
+        task = MathTask(
             variations_per_field=variations_per_field,
             api_platform=api_platform,
             model_name=model_name,
@@ -109,8 +106,8 @@ def generate_sentiment_variations(variations_per_field, api_platform, model_name
             random_seed=random_seed
         )
 
-        # Override output path to save in sentiment folder
-        output_file = output_dir / "sentiment_sst_variations.json"
+        # Override output path to save in math folder
+        output_file = output_dir / "math_gsm8k_variations.json"
 
         # Generate using custom path
         print(f"🚀 Starting {task.task_name}")
@@ -162,16 +159,16 @@ def generate_sentiment_variations(variations_per_field, api_platform, model_name
         print("\n7. Final statistics:")
         task.mp.info()
 
-        print(f"✅ Completed Sentiment Analysis: {output_file}")
+        print(f"✅ Completed Math Problems Task: {output_file}")
         return str(output_file)
 
     except Exception as e:
-        print(f"❌ Error processing sentiment analysis: {e}")
+        print(f"❌ Error processing math problems task: {e}")
         raise
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Generate sentiment analysis prompt variations")
+    parser = argparse.ArgumentParser(description="Generate math problem solving prompt variations")
     parser.add_argument("--rows", type=int, help="Number of rows to process", default=DEFAULT_MAX_ROWS)
     parser.add_argument("--variations", type=int, help="Number of variations per row", default=DEFAULT_MAX_VARIATIONS_PER_ROW)
     parser.add_argument("--variations_per_field", type=int, default=DEFAULT_VARIATIONS_PER_FIELD)
@@ -180,7 +177,7 @@ if __name__ == "__main__":
     parser.add_argument("--random_seed", type=int, help="Random seed for generation", default=DEFAULT_RANDOM_SEED)
     args = parser.parse_args()
 
-    task = SentimentTask(
+    task = MathTask(
         variations_per_field=args.variations_per_field,
         api_platform=args.api_platform,
         model_name=args.model_name,
@@ -190,4 +187,4 @@ if __name__ == "__main__":
     )
     if args.rows is not None or args.variations is not None:
         task.override_config(rows=args.rows, variations=args.variations)
-    task.generate()
+    task.generate()
