+++ promptsuite_tasks/data_generation/mmlu_task.py
@@ -12,7 +12,7 @@ import pandas as pd
 import ast
 from typing import Dict, Any, List
 
-# Add the project root to the path to import multipromptify and multipromptify_tasks
+# Add the project root to the path to import promptsuite and promptsuite_tasks
 project_root = Path(__file__).parent.parent.parent.parent
 sys.path.insert(0, str(project_root))
 
@@ -20,7 +20,7 @@ sys.path.insert(0, str(project_root))
 current_dir = Path(__file__).parent.parent
 sys.path.insert(0, str(current_dir))
 
-from multipromptify.core.template_keys import (
+from promptsuite.core.template_keys import (
     INSTRUCTION, PROMPT_FORMAT, QUESTION_KEY, OPTIONS_KEY, GOLD_KEY,
     PARAPHRASE_WITH_LLM, FORMAT_STRUCTURE_VARIATION, TYPOS_AND_NOISE_VARIATION,
     SHUFFLE_VARIATION, ENUMERATE_VARIATION, INSTRUCTION_VARIATIONS, PROMPT_FORMAT_VARIATIONS,
