+++ promptsuite_tasks/data_generation/qa_task.py
@@ -11,7 +11,7 @@ from typing import Dict, Any
 import pandas as pd
 from datasets import load_dataset
 
-# Add the project root to the path to import multipromptify and multipromptify_tasks
+# Add the project root to the path to import promptsuite and promptsuite_tasks
 project_root = Path(__file__).parent.parent.parent.parent
 sys.path.insert(0, str(project_root))
 
@@ -19,15 +19,15 @@ sys.path.insert(0, str(project_root))
 current_dir = Path(__file__).parent.parent
 sys.path.insert(0, str(current_dir))
 
-from multipromptify.core import FEW_SHOT_KEY
-from multipromptify.core.template_keys import (
+from promptsuite.core import FEW_SHOT_KEY
+from promptsuite.core.template_keys import (
     INSTRUCTION, PROMPT_FORMAT, GOLD_KEY, CONTEXT_KEY,
     PARAPHRASE_WITH_LLM, FORMAT_STRUCTURE_VARIATION, TYPOS_AND_NOISE_VARIATION, INSTRUCTION_VARIATIONS,
     PROMPT_FORMAT_VARIATIONS, CONTEXT_VARIATION
 )
 from base_task import BaseTask
 
-from multipromptify_tasks.constants import (
+from promptsuite_tasks.constants import (
     DEFAULT_VARIATIONS_PER_FIELD, DEFAULT_PLATFORM, DEFAULT_MODEL_NAME,
     DEFAULT_MAX_VARIATIONS_PER_ROW, DEFAULT_MAX_ROWS, DEFAULT_RANDOM_SEED
 )
@@ -71,7 +71,7 @@ class QATask(BaseTask):
         df = pd.concat([train_df, test_df], ignore_index=True)
         print(f"✅ Combined total: {len(df)} rows")
 
-        # Load into multipromptify
+        # Load into promptsuite
         self.mp.load_dataframe(df)
         self.post_process()
         print("✅ Data post-processed")
