+++ promptsuite_tasks/execution/run_musique_batch.py
@@ -15,9 +15,9 @@ import time
 from pathlib import Path
 from typing import Dict, Any, Optional, Callable
 
-from multipromptify_tasks.constants import MODEL_SHORT_NAMES
-from multipromptify_tasks.execution.batch_runner_base import BatchRunnerBase, get_model_name
-from multipromptify_tasks.execution.shared_metrics import calculate_musique_correctness_and_metrics
+from promptsuite_tasks.constants import MODEL_SHORT_NAMES
+from promptsuite_tasks.execution.batch_runner_base import BatchRunnerBase, get_model_name
+from promptsuite_tasks.execution.shared_metrics import calculate_musique_correctness_and_metrics
 
 
 class MuSiQueBatchRunner(BatchRunnerBase):
